% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{delete_backup}
\alias{delete_backup}
\title{Deletes an existing backup of a table}
\usage{
delete_backup(
  BackupArn = "string"
)
}
\arguments{
\item{BackupArn}{[required] The ARN associated with the backup.}
}
\description{
Deletes an existing backup of a table.
}
\details{
You can call \code{DeleteBackup} at a maximum rate of 10 times per second.
}
\examples{

}
