% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{describe_backup}
\alias{describe_backup}
\title{Describes an existing backup of a table}
\usage{
describe_backup(BackupArn)
}
\arguments{
\item{BackupArn}{[required] The ARN associated with the backup.}
}
\description{
Describes an existing backup of a table.
}
\details{
You can call \code{DescribeBackup} at a maximum rate of 10 times per second.
}
\section{Accepted Parameters}{
\preformatted{describe_backup(
  BackupArn = "string"
)
}
}

