% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{describe_continuous_backups}
\alias{describe_continuous_backups}
\title{Checks the status of continuous backups and point in time recovery on the specified table}
\usage{
describe_continuous_backups(TableName)
}
\arguments{
\item{TableName}{Name of the table for which the customer wants to check the continuous backups and point in time recovery settings.}
}
\description{
Checks the status of continuous backups and point in time recovery on the specified table. Continuous backups are \code{ENABLED} on all tables at table creation. If point in time recovery is enabled, \code{PointInTimeRecoveryStatus} will be set to ENABLED.
}
\details{
Once continuous backups and point in time recovery are enabled, you can restore to any point in time within \code{EarliestRestorableDateTime} and \code{LatestRestorableDateTime}.

\code{LatestRestorableDateTime} is typically 5 minutes before the current time. You can restore your table to any point in time during the last 35 days.

You can call \code{DescribeContinuousBackups} at a maximum rate of 10 times per second.
}
\examples{

}
