% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{describe_endpoints}
\alias{describe_endpoints}
\title{Returns the regional endpoint information}
\usage{
describe_endpoints()
}
\description{
Returns the regional endpoint information.
}
\section{Accepted Parameters}{
\preformatted{describe_endpoints()
}
}

