% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{describe_global_table}
\alias{describe_global_table}
\title{Returns information about the specified global table}
\usage{
describe_global_table(GlobalTableName)
}
\arguments{
\item{GlobalTableName}{The name of the global table.}
}
\description{
Returns information about the specified global table.
}
\examples{

}
