% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{describe_global_table_settings}
\alias{describe_global_table_settings}
\title{Describes region specific settings for a global table}
\usage{
describe_global_table_settings(
  GlobalTableName = "string"
)
}
\arguments{
\item{GlobalTableName}{[required] The name of the global table to describe.}
}
\description{
Describes region specific settings for a global table.
}
\examples{

}
