% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{describe_time_to_live}
\alias{describe_time_to_live}
\title{Gives a description of the Time to Live (TTL) status on the specified table}
\usage{
describe_time_to_live(TableName)
}
\arguments{
\item{TableName}{The name of the table to be described.}
}
\description{
Gives a description of the Time to Live (TTL) status on the specified table.
}
\examples{

}
