% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{list_backups}
\alias{list_backups}
\title{List backups associated with an AWS account}
\usage{
list_backups(TableName = NULL, Limit = NULL,
  TimeRangeLowerBound = NULL, TimeRangeUpperBound = NULL,
  ExclusiveStartBackupArn = NULL, BackupType = NULL)
}
\arguments{
\item{TableName}{The backups from the table specified by \code{TableName} are listed.}

\item{Limit}{Maximum number of backups to return at once.}

\item{TimeRangeLowerBound}{Only backups created after this time are listed. \code{TimeRangeLowerBound} is inclusive.}

\item{TimeRangeUpperBound}{Only backups created before this time are listed. \code{TimeRangeUpperBound} is exclusive.}

\item{ExclusiveStartBackupArn}{\code{LastEvaluatedBackupArn} is the ARN of the backup last evaluated when the current page of results was returned, inclusive of the current page of results. This value may be specified as the \code{ExclusiveStartBackupArn} of a new \code{ListBackups} operation in order to fetch the next page of results.}

\item{BackupType}{The backups from the table specified by \code{BackupType} are listed.

Where \code{BackupType} can be:
\itemize{
\item \code{USER} - On-demand backup created by you.
\item \code{SYSTEM} - On-demand backup automatically created by DynamoDB.
\item \code{ALL} - All types of on-demand backups (USER and SYSTEM).
}}
}
\description{
List backups associated with an AWS account. To list backups for a given table, specify \code{TableName}. \code{ListBackups} returns a paginated list of results with at most 1MB worth of items in a page. You can also specify a limit for the maximum number of entries to be returned in a page.
}
\details{
In the request, start time is inclusive but end time is exclusive. Note that these limits are for the time at which the original backup was requested.

You can call \code{ListBackups} a maximum of 5 times per second.
}
\section{Accepted Parameters}{
\preformatted{list_backups(
  TableName = "string",
  Limit = 123,
  TimeRangeLowerBound = as.POSIXct("2015-01-01"),
  TimeRangeUpperBound = as.POSIXct("2015-01-01"),
  ExclusiveStartBackupArn = "string",
  BackupType = "USER"|"SYSTEM"|"AWS_BACKUP"|"ALL"
)
}
}

