% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{list_global_tables}
\alias{list_global_tables}
\title{Lists all global tables that have a replica in the specified region}
\usage{
list_global_tables(ExclusiveStartGlobalTableName = NULL, Limit = NULL,
  RegionName = NULL)
}
\arguments{
\item{ExclusiveStartGlobalTableName}{The first global table name that this operation will evaluate.}

\item{Limit}{The maximum number of table names to return.}

\item{RegionName}{Lists the global tables in a specific region.}
}
\description{
Lists all global tables that have a replica in the specified region.
}
\section{Accepted Parameters}{
\preformatted{list_global_tables(
  ExclusiveStartGlobalTableName = "string",
  Limit = 123,
  RegionName = "string"
)
}
}

