% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_package.R
\docType{package}
\name{paws.dynamodb-package}
\alias{paws.dynamodb}
\alias{paws.dynamodb-package}
\title{paws.dynamodb: Amazon DynamoDB}
\description{
Amazon DynamoDB is a fully managed NoSQL database service
that provides fast and predictable performance with seamless
scalability. DynamoDB lets you offload the administrative burdens of
operating and scaling a distributed database, so that you don\'t have
to worry about hardware provisioning, setup and configuration,
replication, software patching, or cluster scaling.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
