% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{update_global_table_settings}
\alias{update_global_table_settings}
\title{Updates settings for a global table}
\usage{
update_global_table_settings(GlobalTableName,
  GlobalTableProvisionedWriteCapacityUnits = NULL,
  GlobalTableProvisionedWriteCapacityAutoScalingSettingsUpdate = NULL,
  GlobalTableGlobalSecondaryIndexSettingsUpdate = NULL,
  ReplicaSettingsUpdate = NULL)
}
\arguments{
\item{GlobalTableName}{The name of the global table}

\item{GlobalTableProvisionedWriteCapacityUnits}{The maximum number of writes consumed per second before DynamoDB returns a \code{ThrottlingException.}}

\item{GlobalTableProvisionedWriteCapacityAutoScalingSettingsUpdate}{AutoScaling settings for managing provisioned write capacity for the global table.}

\item{GlobalTableGlobalSecondaryIndexSettingsUpdate}{Represents the settings of a global secondary index for a global table that will be modified.}

\item{ReplicaSettingsUpdate}{Represents the settings for a global table in a region that will be modified.}
}
\description{
Updates settings for a global table.
}
\examples{

}
