% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dynamodb_operations.R
\name{describe_table}
\alias{describe_table}
\title{Returns information about the table, including the current status of the table, when it was created, the primary key schema, and any indexes on the table}
\usage{
describe_table(TableName)
}
\arguments{
\item{TableName}{[required] The name of the table to describe.}
}
\description{
Returns information about the table, including the current status of the table, when it was created, the primary key schema, and any indexes on the table.
}
\details{
If you issue a \code{DescribeTable} request immediately after a \code{CreateTable} request, DynamoDB might return a \code{ResourceNotFoundException}. This is because \code{DescribeTable} uses an eventually consistent query, and the metadata for your table might not be available at that moment. Wait for a few seconds, and then try the \code{DescribeTable} request again.
}
\section{Accepted Parameters}{
\preformatted{describe_table(
  TableName = "string"
)
}
}

\examples{
# This example describes the Music table.
\donttest{describe_table(
  TableName = "Music"
)}

}
